Add module to compare TPM logs
The difference in 'old good' and 'new' TPM logs is not very big so LCS
algorithm works just fine with one exception: if events are reordered
they are detected as removed and added. To mitigate this 'semantic'
comparison is implemented e.g. BootOrder [1,2,3] and BootOrder [2,1,3]
are the same events with modified data so the event is detected as
'modified'

Signed-off-by: Mikhail Malyshev <mike.malyshev@gmail.com>